chore: add simpler way to create metadata
